finished  rest api with mongoose

+++ controllers/feed.js
@@ -2,13 +2,28 @@ const fs = require('fs');
 const path = require('path');
 const { validationResult } = require('express-validator/check');
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
-    Post.find().then(posts => {
+    const currentPage = req.query.page || 1;
+    const perPage = 2;
+    let totalItems;
+    Post.find()
+    .countDocuments()
+    .then(count => {
+        totalItems = count;
+        return Post.find()
+        .skip((currentPage-1) * perPage)
+        .limit(perPage);
+    })
+    .then(posts => {
         res.status(200).json({
-            message: 'fetched post successfully.', posts: posts
+            message: 'fetched post successfully.',
+             posts: posts, 
+             totalItems: totalItems
         });
-    }).catch(err => {
+    })
+    .catch(err => {
         if(!err.statusCode){
             err.statusCode = 500;
         }
@@ -36,10 +51,6 @@ exports.createPost = (req, res, next) => {
         const error = new Error('Validation failed, entered data is incorrect.');
         error.statusCode = 422;
         throw error;
-        // return res.status(422).json({
-        //     message: 'Validation failed, entered data is incorrect.',
-        //     errors: errors.array()
-        // })
     }
     if(!req.file){
         const error = new Error('No image provided.');
@@ -51,18 +62,32 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\", "/"); // use for only window
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
          content: content,
          imageUrl: imageUrl,
-          creator: {name: 'PingPong'}
+          creator: req.userId
     });
-    post.save().then(result => {
+    post.save()
+    .then(result => {
+        return User.findById(req.userId);
+       
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+        
+    })
+    .then(result => {
         res.status(201).json({
             message: 'Post created successfully!',
-            post: result
+            post: post,
+            creator: {_id: creator._id, name: creator.name}
         });
-    }).catch(err=>{
+    })
+    .catch(err=>{
         if(!err.statusCode){
             err.statusCode = 500;
         }
@@ -115,6 +140,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
         }
+        if(post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized!');
+            error.statusCode = 404;
+            throw error;
+        }
         if(imageUrl !== post.imageUrl){
             clearImage(post.imageUrl);
         }
@@ -148,12 +178,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
         }
+        if(post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized!');
+            error.statusCode = 404;
+            throw error;
+        }
         // Check logged in user
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-        console.log(result);
+       return User.findById(req.userId);
+    })
+    .then(user => {
+        user.posts.pull(postId);
+        return user.save();
+    })
+    .then(result => {
         res.status(200).json({message: 'Deleted post.'});
     })
     .catch(err => {
